Add all serialzers ref_name

+++ apps/event/api/v1/serializers/event_serializer.py
@@ -8,42 +8,47 @@ class ImageSerializer(serializers.ModelSerializer):
     class Meta:
         model = EventImage
         fields = "__all__"
+        ref_name = "EventImage"
 
     def get_image_url(self, obj):
         return obj.image.url if obj.image else None
 
 
-class EventSerializer(serializers.ModelSerializer):
+class EventReadSerializer(serializers.ModelSerializer):
     images = ImageSerializer(many=True, read_only=True)
 
     class Meta:
         model = Event
         fields = ("id", "title", "organizer", "email", "phone", "more", "date", "address", "images")
+        ref_name = "EventRead"
 
 
-class EventCreateUpdateSerializer(serializers.ModelSerializer):
-    images = ImageSerializer(many=True, read_only=True)
+class EventWriteSerializer(serializers.ModelSerializer):
+    images = serializers.ListField(child=serializers.ImageField(), write_only=True, required=False)
 
     class Meta:
         model = Event
         fields = ("id", "title", "organizer", "email", "phone", "more", "date", "address", "images")
+        ref_name = "EventWrite"
 
-    def create(self, validated_data):
+    def save_images(self, instance, is_update=False):
         request = self.context.get("request")
         images_data = request.FILES.getlist("images", [])
-        event = Event.objects.create(**validated_data)
+
+        if is_update:
+            instance.images.all().delete()
+
         for image_data in images_data:
-            EventImage.objects.create(event=event, image=image_data)
+            EventImage.objects.create(event=instance, image=image_data)
+
+    def create(self, validated_data):
+        event = Event.objects.create(**validated_data)
+        self.save_images(event)
         return event
 
     def update(self, instance, validated_data):
-        request = self.context.get("request")
-        images_data = request.FILES.getlist("images", [])
         instance = super().update(instance, validated_data)
-
-        instance.images.all().delete()
-        for image_data in images_data:
-            EventImage.objects.create(event=instance, image=image_data)
+        self.save_images(instance, is_update=True)
         return instance
 
     def validate(self, data):
